# UI: Optimize treeview render/filter performance