--- a/examples/cra-kitchen-sink/src/stories/perf.stories.js
+++ b/examples/cra-kitchen-sink/src/stories/perf.stories.js
@@ -4,13 +4,13 @@ import { storiesOf } from '@storybook/react';
 
 import { Button } from '@storybook/react/demo';
 
-for (let i = 0; i < 0; i += 1) {
+for (let i = 0; i < 1; i += 1) {
   const randomDemoName = Math.random()
     .toString(36)
     .substring(7);
-  const stories = storiesOf(`Perf|${i}/${randomDemoName}`);
+  const stories = storiesOf(`Perf.${randomDemoName}`);
 
-  for (let j = 0; j < 100; j += 1) {
+  for (let j = 0; j < 10; j += 1) {
     stories
       .add(`with text ${j}`, () => <Button>Hello Button</Button>)
       .add(`with emoji ${j}`, () => (