--- a/lib/ui/src/components/sidebar/treeview/treeview.js
+++ b/lib/ui/src/components/sidebar/treeview/treeview.js
@@ -150,14 +150,10 @@ const calculateTreeState = memoize(50)(
 
     // If a new selection is made, we need to ensure it is part of the expanded set
     const selectedAncestorIds = selectedId ? getParents(selectedId, dataset).map(i => i.id) : [];
-
-    const newExpanded = Object.keys(dataset).reduce(
-      (acc, key) => ({
-        ...acc,
-        [key]: selectedAncestorIds.includes(key) || unfilteredExpanded[key],
-      }),
-      {}
-    );
+    const newExpanded = Object.keys(dataset).reduce((acc, key) => {
+      acc[key] = selectedAncestorIds.includes(key) || unfilteredExpanded[key];
+      return acc;
+    }, {});
 
     return {
       lastSelectedId: selectedId,