--- a/lib/ui/src/components/sidebar/treeview/utils.js
+++ b/lib/ui/src/components/sidebar/treeview/utils.js
@@ -172,20 +172,17 @@ export const toFiltered = (dataset, filter) => {
 
   // get all parents for all results
   const result = found.reduce((acc, item) => {
-    const parents = getParents(item.id, dataset).reduce(
-      (pacc, pitem) => ({ ...pacc, [pitem.id]: { ...pitem } }),
-      {}
-    );
+    getParents(item.id, dataset).forEach(pitem => {
+      acc[pitem.id] = pitem;
+    });
 
-    return { ...acc, [item.id]: item, ...parents };
+    acc[item.id] = item;
+    return acc;
   }, {});
 
   // filter the children of the found items (and their parents) so only found entries are present
-  return Object.entries(result).reduce(
-    (acc, [k, v]) => ({
-      ...acc,
-      [k]: v.children ? { ...v, children: v.children.filter(c => !!result[c]) } : v,
-    }),
-    {}
-  );
+  return Object.entries(result).reduce((acc, [k, v]) => {
+    acc[k] = v.children ? { ...v, children: v.children.filter(c => !!result[c]) } : v;
+    return acc;
+  }, {});
 };