Issue: #7907

## What I did

- Recreated a simplified version of @kevinwebber's perf test inside `cra-kitchen-sink`
- Optimized a few key loops in the treeview UI (w/ 2000 extra `perf` stories)
  - calculateTreeState (render):
    - before: ~750ms
    - after: ~10ms
  - toFiltered (search):
    - before: ~800ms
    - after: ~50ms

## How to test

Edit `perf.stories.js` to set the outer loop max to 10 & inner to 100. Then:

```
cd examples/cra-kitchen-sink
yarn storybook --no-dll
```