@kevinweber 2019-08-30
Wow, I love this, such a great fix!

PS: @shilman You tagged the wrong Kevin Weber in the PR description. It should be @kevinweber (with one `b`)