@theKashey 2019-09-17
What about increasing the limit for `getParents`/`getParent`? Nowadays it's just 1000 elements, and using `array.instead` instead of _spread_ at `getParents`?

Now sure about the impact, but also a low hanging fruit.